# docs/api/openapi/orchestrator.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentBox Orchestrator",
    "description": "Sandbox routing, worker management, and admin API",
    "version": "0.1.0"
  },
  "paths": {
    "/internal/workers/register": {
      "post": {
        "tags": [
          "workers"
        ],
        "summary": "Register Worker",
        "description": "Worker calls this on startup to join the pool.",
        "operationId": "register_worker_internal_workers_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/workers/heartbeat": {
      "post": {
        "tags": [
          "workers"
        ],
        "summary": "Worker Heartbeat",
        "description": "Worker calls this periodically to update capacity and stay alive.",
        "operationId": "worker_heartbeat_internal_workers_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerHeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/workers/deregister": {
      "post": {
        "tags": [
          "workers"
        ],
        "summary": "Deregister Worker",
        "description": "Worker calls this on graceful shutdown (SIGTERM).",
        "operationId": "deregister_worker_internal_workers_deregister_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerHeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workers": {
      "get": {
        "tags": [
          "workers"
        ],
        "summary": "List Workers",
        "description": "List all registered workers with capacity info.",
        "operationId": "list_workers_workers_get",
        "parameters": [
          {
            "name": "state_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkerHeartbeatRequest": {
        "properties": {
          "worker_id": {
            "type": "string",
            "title": "Worker Id"
          },
          "active_sandboxes": {
            "type": "integer",
            "title": "Active Sandboxes",
            "default": 0
          },
          "state": {
            "type": "string",
            "title": "State",
            "default": "active"
          }
        },
        "type": "object",
        "required": [
          "worker_id"
        ],
        "title": "WorkerHeartbeatRequest"
      },
      "WorkerRegisterRequest": {
        "properties": {
          "worker_id": {
            "type": "string",
            "title": "Worker Id"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "max_sandboxes": {
            "type": "integer",
            "title": "Max Sandboxes",
            "default": 50
          },
          "active_sandboxes": {
            "type": "integer",
            "title": "Active Sandboxes",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "worker_id",
          "endpoint"
        ],
        "title": "WorkerRegisterRequest"
      },
...
    }
  }
}